Atharva3

+++ src/main/java/com/capgemini/flightbookingsystem/services/BookingService.java
@@ -8,12 +8,12 @@ public interface BookingService {
 
 	Booking saveBooking(Booking booking);
 
-	Booking getBookingById(Long bookingId);
+	Booking getBookingById(Integer bookingId);
 
 	List<Booking> getAllBookings();
 
-	Booking updateBooking(Long bookingId, Booking booking);
+	Booking updateBooking(Integer bookingId, Booking booking);
 
-	boolean deleteBooking(Long bookingId);
+	boolean deleteBooking(Integer bookingId);
 
 }
